\name{simulatorBRW}
\alias{simulatorBRW}
\title{
A function to simulate a Bownian random walk (BRW) diffusion process
}
\description{
This function simulates a Bownian random walk (BRW) diffusion process while preventing dispersal in inaccessible raster cells.
}
\usage{
simulatorBRW(envVariable =
raster(matrix(c(runif(600,5,10),runif(1000,0,5)),nrow=40,ncol=40)),
			  d = 1,
			  ancestPosition = c(0.4,0.5),
			  birthRate = 0.2,
			  samplingRate = 0.2,
			  startingYear = 0,
			  samplingWindow = c(10,50),
			  timeSlice = 0.1, 
			  timeIntervale = 1,
			  showingPlots = FALSE,
			  extractionOfValuesOnMatrix = FALSE)
}
\arguments{
  \item{envVariable}{ raster that will determine the environmental heterogeneity (cells with a NaN value will be considered as inaccessible). }
  \item{d}{ distance, in decimal degrees, travelled by each lineage during each time slice. }
  \item{ancestPosition}{ vector of geographic coordinates (longitude, latitude) of the most ancestral node position, i.e. the starting position of the RRW diffusion simulation. }
  \item{birthRate}{ value defining the rate at which lineages are splited in two new lineages (expressed in events per lineage per time unit). }
  \item{samplingRate}{ value defining the rate at which lineage are sampled and thus stopped moving and splitting on the raster (expressed in events per lineage per time unit). }
  \item{startingYear}{ year (or moment in another time unit) at which the simulation begins from the ancestral node position. }
  \item{samplingWindow}{ vector of time values (in the same time unit as the "startingYear") defining the starting and ending times of the sampling period, i.e. the period during which the sampling rate will be effective. }
  \item{timeSlice}{ time interval (in the same time unit as the "startingYear") at which unsampled lineages perform a new movement on the raster. }
  \item{timeIntervale}{ time interval (in the same time unit as the "startingYear") at which the plot displaying the simulation is updated (only taken into account when showingPlots=TRUE). }
  \item{showingPlots}{ boolean variable specifying if the different plots have to be displayed or not. }
  \item{extractionOfValuesOnMatrix}{ boolean variable specifying if the raster object as to be preliminary transformed into a matrix object. This operation allows a significantly faster simulation but has to be avoided if possible. Indeed, with this transformation, the raster and the ancestral position are projected on a flat grid and thus lose their initial projection during the whole simulation. This aspect can, in some cases, lead to over-simplistic situations. Note that if this option is selected, coordinates of simulated nodes are re-projected on the initial raster at the end of the simulation. }
}
\value{
The function simulates a RRW diffusion process based on an heterogeneous environmental raster. In practice, the function generates a list of two objects: (i) a table containing the spatio-tremporal information of the simulated RRW diffusion, with one line per lineage/phylogenetic branch. Such a table similar to those generated by the "treeExtractions" function. (ii) A tree object of class "phylo" that thus summarizes the phylogenetic relationships among simulated lineages.
}
\note{
In order to avoid null standard deviation values, all the raster cell values are preliminary increased by "1" (see the formulas below). Furthermore, if the environmental raster is treated as a resistance factor, all its cell values will be inverted (after having increased all the cell values by one). This operation allows using the same standard deviation formulas for both the cases where the environmental raster as to be treated as a conductance or a resistance factor. \cr
\cr
For a given raster cell \eqn{i} with environmental \eqn{v_i}, standard deviations (\eqn{SDx_i} and \eqn{SDy_i}) used to define the two normal distributions with mean = 0, in which longitudinal \eqn{d_x} and latitudinal \eqn{d_y} displacement values are randomly picked, are computed as follows: \cr
\cr
\eqn{SDx_i = (v_i/v_{max}) * cell_{width} * scalingValue} \cr
\eqn{SDy_i = (v_i/v_{max}) * cell_{height} * scalingValue} \cr
\cr
with \eqn{v_{max}} being the highest value recorded in the environmental raster. Note that \eqn{cell_{width}} and \eqn{cell_{height}} are taken from the raster properties/parameters.
Final note: in order to obtain to simulate trees with a realistic sampling pattern through time, tip branches that have not been sampled at the end of simulation are removed from the simulated tree topology.}
\references{
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
}
\author{
Simon Dellicour
}
