\name{simulatorRRW1}
\alias{simulatorRRW1}
\title{
A function to simulate relaxed random walk (RRW) diffusion along a phylogenetic tree
}
\description{
This function simulates relaxed random walk (RRW) diffusion based on parameters previously estimated with the continuous phylogeographic model implemented in BEAST (Lemey \emph{et al}. 2010).
}
\usage{
simulatorRRW1(tree,
             rates,
             sigmas = c(0.1,0.1),
             cor = 0,
             envVariables = list(),
             mostRecentSamplingDatum, 
             ancestPosition = c(0,0),
             reciprocalRates = TRUE,
             n1 = 100,
             n2 = 0,
             showingPlots = FALSE,
             newPlot = TRUE)                
}
\arguments{
  \item{tree}{ phylogenetic tree (object of class "phylo"). }
  \item{rates}{ vector of trait evolutionary rates estimated by BEAST and associated with each branch (vector elements have to be in the same order as their associated branch edges in the "tree" object. }
  \item{sigmas}{ vector of scale parameters for the latitude and longitude (can be retrieved from BEAST output; see the related tutorial for further details). }
  \item{cor}{ observed correlation between latitude and longitude (can be retrieved from BEAST output; see the related tutorial for further details). }
  \item{envVariables}{ list of environmental rasters that will determine the accessible area (simulated node positions will not fall into raster cells with NaN values). }
  \item{mostRecentSamplingDatum}{ most recent datum of sampling in decimal format (e.g. 2007.4). }
  \item{ancestPosition}{ vector of geographic coordinates (longitude, latitude) of the most ancestral node position, i.e. the starting position of the (R)RW simulation. }
  \item{reciprocalRates}{ BEAST parameter of the continuous diffusion model: boolean variable specifying if branch lengths have to divided (TRUE) or divided (FALSE) by the trait evolutionary rates. }
  \item{n1}{ maximum number of branch rotations allowed per simulated node position. If a simulated node position falls into an inaccessible area (i.e. a raster cell with a NaN value), the simulator can rotate the branch around its oldest node position and while maintaining the geographic distance travelled by the branch until its youngest node position falls into an accessible area. In that situation, "n1" defines the maximum number of times that this rotation trial can be attempted. After "n1" rotation trials, the branch is not rotated anymore and the youngest node position (and hence its actual distance from the oldest node of the branch) is re-simulated if "n2" is higher than "0" (see below) or the entire RRW simulation is restarted from the root of the tree if n2 is set to "0". }
  \item{n2}{ maximum number of re-simulations allowed per node position. If a simulated node position falls into an inaccessible area (i.e. a raster cell with a NaN value), the simulator will first attempt to rotate the branch around its oldest node position until its youngest node position falls into an accessible area. But after "n1" rotation trials, and if "n2" is higher than "0", it will re-simulate the youngest node position and hence its actual distance from the oldest node of the branch. In that situation, "n2" defines the maximum number of times that a new position is re-simulated (and then also rotated up to n1 times) before restarting the entire RRW simulation from the root of the tree. }
  \item{showingPlots}{ boolean variable specifying if the different plots have to be displayed or not. }
  \item{newPlot}{ boolean variable specifying if a new plot window has to be opened or if the simulated branches/nodes have to be plotted on a previously opened plot window e.g. with a specific raster and/or polygons already mapped on it (graphical parameter that is only useful when showingPlots=TRUE). }
}
\value{
The function simulates RRW diffusion process from a specified ancestral location and along a given phylogenetic tree. In practice, the function generates a table containing the spatio-tremporal information of the simulated RRW diffusion. This table are similar to those generated by the "treeExtractions" function.
}
\author{
Simon Dellicour, Paul Bastide, Philippe Lemey
}
\references{
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
Lemey P, Rambaut A, Welch JJ, Suchard MA (2010). Phylogeography takes a relaxed random walk in continuous space and time. \emph{Molecular Biology & Evolution} 27: 1877-1885.
}
