\name{treesRandomisation}
\alias{treesRandomisation}
\title{
A function to randomise tree Branches
}
\description{
This function randomises phylogenetic branches on the study area and is based "extraction" tables previously generated by the "treeExtractions" function which summarises each phylogenetic branch as a movement vector. Four distinct randomisations have been implemented (see below).
}
\usage{
treesRandomisation(localTreesDirectory = "",
              nberOfExtractionFiles = 1,
              envVariables=list(),
              randomProcedure = 3)
}
\arguments{
  \item{localTreesDirectory}{ directory where the extracted matrices are saved (see the "treeExctractions" function). }
  \item{nberOfExtractionFiles}{ number of files created with the "treeExtractions" function. }
  \item{envVariables}{ list of environmental rasters to consider. These rasters are only use to prevent that randomised node positions fall in non-accessible areas, i.e. raster cells associated with "NA" value (e.g. sea areas). }
  \item{randomProcedure}{ randomisation procedure to use: "3" (randomisations of nodes position while maintaining the branches length, the tree topology and the location of the most ancestral node), "4" (randomisations of nodes position while maintaining the branches length and the location of the starting node of each branch), "5" (randomisations of nodes position while maintaining the branches length and the location of the ending node of each branch), or "6" (randomisations of nodes position while maintaining only the branches length). }
}
\value{
In practice, the function generates a table containing the spatio-tremporal information for each randomised tree. This table are similar to those generated by the "treeExtractions" function.  
}
\details{
Several randomisation procedures of tree branches have been implemented: (1) a randomisation of the branches position while maintaining their length, the tree topology and the location of the most ancestral node (procedure n°3), (2) a randomisation of the branches position while maintaining their length and the location of the starting node of each branch (procedure n°4), and (3) a randomisation of the branches position while maintaining only their length  (procedure n°5). The latter procedure can be useful, e.g., when analyzing non phylogenetically informed movement data like capture-mark-recapture movement data. Note that in each case, branch positions are independently randomised for each tree.
}
\author{
Simon Dellicour
}
\references{
Dellicour S, Rose R, Pybus OG (2016). Explaining the geographic spread of emerging epidemics: a framework for comparing viral phylogenies and environmental landscape data. \emph{BMC Bioinformatics} 17: 82. \cr \cr
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
}
